feat: Complete a series of feature iterations and bug fixes
This commit includes a large number of feature updates and issue fixes:
1. Add HTTP authentication failure and excessive redirect count error types
2. Support `bt-detach-seed-only` configuration option, automatically enter seeding mode upon completion
3. Fix FTP PASV response parsing vulnerability, add port and IP validity checks
4. Implement SFTP resume support and length verification
5. Optimize HTTP connection manager, support shared cookie storage
6. Add LPD broadcast manager interface, support multicast interface configuration
7. Fix Metalink download processing logic, support pure torrent metalink downloads
8. Optimize disk write flow, add `sync_all` to ensure data persistence
9. Improve session persistence functionality, support Netscape format cookie storage
10. Fix multiple concurrent download and resource management detail issues

+++ aria2-core/src/ftp/connection/negotiation/parsing.rs
@@ -97,9 +97,12 @@ pub(super) fn parse_pasv_response(response: &str) -> Option<(String, u16)> {
     let h2: u8 = parts[1].trim().parse().ok()?;
     let h3: u8 = parts[2].trim().parse().ok()?;
     let h4: u8 = parts[3].trim().parse().ok()?;
-    let p1: u16 = parts[4].trim().parse().ok()?;
-    let p2: u16 = parts[5].trim().parse().ok()?;
-    Some((format!("{}.{}.{}.{}", h1, h2, h3, h4), p1 * 256 + p2))
+    let p1: u8 = parts[4].trim().parse().ok()?;
+    let p2: u8 = parts[5].trim().parse().ok()?;
+    Some((
+        format!("{}.{}.{}.{}", h1, h2, h3, h4),
+        u16::from(p1) * 256 + u16::from(p2),
+    ))
 }
 
 /// Parse EPSV response to extract port.
